% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{image_cel}
\alias{image_cel}
\title{inclusao da logo do cel em imagens}
\usage{
image_cel(endereco, escala_image = "2000", salvar = TRUE,
  nome = "defualt", logo = NULL, escala_logo = "200x200")
}
\arguments{
\item{endereco}{endereco da imagem que deseja incluir a logo do cel}

\item{escala_image}{por padadrao tem 2000px, escala_image = "2000". exemplo por dimensao escala_image = "400X400"}

\item{salvar}{logical, se TRUE salva o aquivo no diretorio.}

\item{nome}{nome do arquivo, se nao for escolhido sera salvo como nova_image_data.png}

\item{logo}{por padrao é usada a logo do CEL, mas pode ser substituido por outra imagem}

\item{escala_logo}{por padrao é "200x200"}
}
\description{
inclusao da logo do cel em imagens
}
\examples{
\dontrun{
image_cel("foto.jpg")
}
}
